+++ rest_framework/__init__.py
@@ -8,7 +8,7 @@ r"""
          __/ |
         |___/
 """
-VERSION = (0, 2, 0)
+VERSION = (0, 2, 2)
 
 __title__ = 'Python-Rest-Framework'
 __author__ = 'Deys Timofey'
@@ -17,5 +17,3 @@ __copyright__ = 'Copyright (c) 2018 Deys Timofey'
 __license__ = 'Apache License 2.0'
 __url__ = 'https://github.com/nxexox/python-rest-framework'
 __version__ = '.'.join(map(str, VERSION))
-
-# Auto version https://github.com/pypa/setuptools_scm